pyproject.toml
[tool.poetry]
name = "fast-zero"
version = "0.1.0"
description = ""
authors = ["Elysson MR <elyssonmr@gmail.com>"]
readme = "README.md"
packages = [{include = "fast_zero"}]

[tool.poetry.dependencies]
python = "3.11.*"
fastapi = "^0.111.0"
ipython = "^8.25.0"
sqlalchemy = "^2.0.31"
alembic = "^1.13.2"
pydantic-settings = "^2.3.4"
pyjwt = "^2.8.0"
pwdlib = {extras = ["argon2"], version = "^0.2.0"}
python-multipart = "^0.0.9"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
taskipy = "^1.12.2"
ruff = "^0.4.9"
httpx = "^0.27.0"
ignr = "^2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'
run = 'fastapi dev fast_zero/app.py'
pre_test = 'task format'
test = 'pytest -s --cov=fast_zero --cov-report term-missing -vv'
post_test = 'coverage html'
init_migration = 'alembic init migrations'
migrate = 'alembic upgrade head'
